rename subfolder mbed-util to core-util

+++ core-util/PoolAllocator.h
@@ -21,7 +21,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-// [TODO] this should probably be 8, but at the moment our allocators are 
+// [TODO] this should probably be 8, but at the moment our allocators are
 // aligned at 4 bytes
 // [TODO] where should the system allocator alignment be defined?
 #define MBED_UTIL_POOL_ALLOC_DEFAULT_ALIGN       4
@@ -45,12 +45,12 @@ public:
                to satisfy the 'alignment' argument)
       * @param alignment allocation alignment in bytes (must be a power of 2, at least 4)
       */
-    PoolAllocator(void *start, size_t elements, size_t element_size, unsigned alignment = MBED_UTIL_POOL_ALLOC_DEFAULT_ALIGN);
+    PoolAllocator(void *start, size_t elements, size_t element_size, size_t alignment = MBED_UTIL_POOL_ALLOC_DEFAULT_ALIGN);
 
     /** Allocate a new element from the pool
      * @returns the address of the new element or NULL for error
-      */
-    void* alloc();
+     */
+    void *alloc();
 
     /** Allocate a new element from the pool and initialize it with 0
       * @returns the address of the new element or NULL for error
